requirements.txt
Django
pymongo
mongoengine
django-cors-headers
djangorestframework
djangorestframework-simplejwt==5.3.1 
pyjwt==2.10.1
python-dotenv==1.0.1
django_extensions==3.2.3
gunicorn==20.1.0
WeasyPrint
google-generativeai